Extract #versions and #fetched? into Henson::Source::Tarball
