+++ lib/henson/source/tarball.rb
@@ -1,0 +1,23 @@
+require "pathname"
+require "rubygems/package"
+
+module Henson
+  module Source
+    class Tarball < Generic
+
+      # Public: Check if the module tarball has been cached.
+      #
+      # Returns True if the module tarball exists on disk, otherwise False.
+      def fetched?
+        cache_path.file?
+      end
+
+      # Public: Determine the version of the module to be installed.
+      #
+      # Returns the String version number.
+      def version
+        @version ||= resolve_version_from_requirement(@requirement)
+      end
+    end
+  end
+end
